svm: move PhysicalMemory#getUsedSizeFromProcMemInfo to LinuxPhysicalMemorySupportImpl
