+++ substratevm/src/com.oracle.svm.core.posix/src/com/oracle/svm/core/posix/linux/LinuxPhysicalMemorySupportImpl.java
@@ -24,16 +24,27 @@
  */
 package com.oracle.svm.core.posix.linux;
 
+import java.io.BufferedReader;
+import java.io.FileReader;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.List;
+
 import org.graalvm.word.UnsignedWord;
 import org.graalvm.word.WordFactory;
 
+import com.oracle.svm.core.container.Container;
+import com.oracle.svm.core.feature.AutomaticallyRegisteredImageSingleton;
+import com.oracle.svm.core.heap.PhysicalMemory;
 import com.oracle.svm.core.heap.PhysicalMemory.PhysicalMemorySupport;
 import com.oracle.svm.core.posix.headers.Unistd;
-import com.oracle.svm.core.feature.AutomaticallyRegisteredImageSingleton;
 import com.oracle.svm.core.util.VMError;
 
 @AutomaticallyRegisteredImageSingleton(PhysicalMemorySupport.class)
 class LinuxPhysicalMemorySupportImpl implements PhysicalMemorySupport {
+
+    private static final long K = 1024;
 
     @Override
     public UnsignedWord size() {
@@ -43,5 +54,64 @@ class LinuxPhysicalMemorySupportImpl implements PhysicalMemorySupport {
             throw VMError.shouldNotReachHere("Physical memory size (number of pages or page size) not available");
         }
         return WordFactory.unsigned(numberOfPhysicalMemoryPages).multiply(WordFactory.unsigned(sizeOfAPhysicalMemoryPage));
+    }
+
+    @Override
+    public long usedSize() {
+        /*
+         * Note: we use getCachedMemoryLimitInBytes() because we don't want to mutate the state, and
+         * we assume that the memory limits have be queried before calling this method.
+         */
+        assert !(Container.singleton().isContainerized() && Container.singleton().getCachedMemoryLimitInBytes() > 0) : "Should be using OperatingSystemMXBean";
+        /* Non-containerized Linux uses /proc/meminfo. */
+        return getUsedSizeFromProcMemInfo();
+    }
+
+    private static long getUsedSizeFromProcMemInfo() {
+        try {
+            List<String> lines = readAllLines("/proc/meminfo");
+            for (String line : lines) {
+                if (line.contains("MemAvailable")) {
+                    return PhysicalMemory.getCachedSize().rawValue() - parseFirstNumber(line) * K;
+                }
+            }
+        } catch (Exception e) {
+            /* Nothing to do. */
+        }
+        return -1L;
+    }
+
+    private static List<String> readAllLines(String fileName) throws IOException {
+        List<String> lines = new ArrayList<>();
+        try (BufferedReader bufferedReader = new BufferedReader(new FileReader(fileName, StandardCharsets.UTF_8))) {
+            String line;
+            while ((line = bufferedReader.readLine()) != null) {
+                lines.add(line);
+            }
+        }
+        return lines;
+    }
+
+    /** Parses the first number in the String as a long value. */
+    private static long parseFirstNumber(String str) {
+        int firstDigit = -1;
+        int lastDigit = -1;
+
+        for (int i = 0; i < str.length(); i++) {
+            if (Character.isDigit(str.charAt(i))) {
+                if (firstDigit == -1) {
+                    firstDigit = i;
+                }
+                lastDigit = i;
+            } else if (firstDigit != -1) {
+                break;
+            }
+        }
+
+        if (firstDigit >= 0) {
+            String number = str.substring(firstDigit, lastDigit + 1);
+            return Long.parseLong(number);
+        }
+        return -1;
     }
 }
